Add redirect to a connection specified in BUILDCONFIG
Allow creating tokens on redirected connection
